+++ components/forms/LoginForm.vue
@@ -1,8 +1,4 @@
 <script setup lang="ts">
-defineProps({
-  register: Boolean,
-})
-
 import { useForm } from 'vee-validate'
 import { toTypedSchema } from '@vee-validate/zod'
 import * as z from 'zod'
@@ -14,34 +10,52 @@ const schema = z.object({
   email: z.string().email(),
   password: z.string().min(8),
 })
+
 const form = useForm({
   validationSchema: toTypedSchema(schema),
 })
 
-async function onSubmit(values: Record<string, any>) {
-  const error = await login('credentials', values)
-
-  if (error) {
-    toast.error(error)
+const onSubmit = form.handleSubmit(
+  async (values) => {
+    const error = await login('credentials', {
+      email: values.email,
+      password: values.password,
+    })
+    if (error) {
+      toast.error(error)
+    }
+  },
+  () => {
+    toast.error('Please fix the errors in the form.')
   }
-}
+)
 </script>
 
 <template>
-  <AutoForm
-    class="space-y-4"
-    :form="form"
-    :schema="schema"
-    :field-config="{
-      email: { inputProps: { type: 'email' } },
-      password: { inputProps: { type: 'password' } },
-    }"
-    @submit="onSubmit"
-  >
+  <form @submit="onSubmit">
+    <FormField v-slot="{ componentField }" name="email">
+      <FormItem>
+        <FormLabel>Email</FormLabel>
+        <FormControl>
+          <Input v-bind="componentField" />
+        </FormControl>
+        <FormDescription />
+        <FormMessage />
+      </FormItem>
+    </FormField>
+    <FormField v-slot="{ componentField }" name="password">
+      <FormItem>
+        <FormLabel>Password</FormLabel>
+        <FormControl>
+          <Input v-bind="componentField" type="password" />
+        </FormControl>
+        <FormDescription />
+        <FormMessage />
+      </FormItem>
+    </FormField>
     <TermsInfo />
-
     <div class="flex justify-end">
       <Button type="submit"> Login </Button>
     </div>
-  </AutoForm>
+  </form>
 </template>
